+++ CHANGELOG.md
@@ -66,6 +66,7 @@
 - Removed `shrink_dtype` method from `Expr` (use `Series#shrink_dtype` instead)
 - Removed `replace` method from `DataFrame` (use `[]=` instead)
 - Removed `fetch` method from `LazyFrame` (use `head` + `collect` instead)
+- Removed `with_context` method from `LazyFrame` (use `Polars.concat(how: "horizontal")` instead)
 - Removed `datetime` method from `DateTimeExpr` (use `replace_time_zone` instead)
 - Removed `mean` and `median` methods from `DateTimeNameSpace` (use `Series#mean` and `Series#median` instead)
 - Removed `concat` method from `StringExpr` and `StringNameSpace` (use `join` instead)